requirements.txt
# requirements.txt
python-dotenv
selenium
notion-client
openpyxl
webdriver-manager